+++ migrations/000192_clusters_org_id_not_null.down.sql
@@ -1,0 +1,2 @@
+ALTER TABLE clusters
+    ALTER COLUMN org_id DROP NOT NULL;